pyproject.toml
[tool.poetry]
name = "AdventOfCode"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.3"
pyhamcrest = "^2.1.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
